RN iOS: Fix test harness
Changes:
 - package.json: detox 12.1.1 -> detox 18.9.0
 - examples/TestDriver/package.json: detox 12.1.1 -> detox 18.9.0

Explanation:
npm install would fail in the root directory and in examples/TestDriver due to an issue with XCode 12.x not compiling the EarlGrey dependency (see https://github.com/google/EarlGrey/issues/1461). Apparently, an improperly escaped backspace character in a comment in a header file was causing XCode to error out. Google merged a fix for this on October 23, 2020, but it looks like Detox actual forks EarlGrey (see https://github.com/wix-playground/EarlGrey), and by the time the fix was merged by Google, Detox was already on version 17.10.3 and its version of the affected file was different. To settle the matter once and for all, I decided to upgrade Detox from 12.1.1 to the latest version, 18.9.0.

refs HEAP-20926