RN iOS: Fix test harness
Changes:
 - examples/TestDriver/React.podspec.patch: Patch file to apply to examples/TestDriver/node_modules/react-native/React.podspec after running `npm install` and before running `pod install`.
 - examples/TestDriver/README.md: Added instructions for how to build TestDriver app.

Explanation:
When attempting to build the TestDriver project for iOS, an error would occur at linking time. Certain symbols from glog and DoubleConversion would be referenced in the cxxreact module, but the linking command would not include those frameworks in the invocation. The reason was that the React.podspec file failed to specify those frameworks as dependencies for cxxreact. The bug was ultimately fixed in RN 0.60.0. However, I would like to keep this app on RN 0.57.0, and so I created a patch based on the fix that was put into 0.60.0. (See https://github.com/facebook/react-native/pull/23430 and https://github.com/facebook/react-native/issues/23390.)

refs HEAP-20926