RN iOS: Fix test harness
Changes:
 - examples/TestDriver/ios/Podfile: Added logic to remove arm64 architecture from iPhone simulator builds.
 - examples/TestDriver/ios/TestDriver.xcodeproj/project.pbxproj: Removed arm64 architecture from debug builds for TestDriver and TestDriverTests build configurations.

Explanation:
As part of the move toward Apple Silicon, Xcode 12 now considers arm64 a valid architecture for simulators. Of course, simulators just run whatever architecture the host machine runs. When I attempt to build the TestDriver app for the simulator on my machine, I see the following error:

ld: in /Users/andrewyousef/react-native-heap/examples/TestDriver/node_modules/@heap/react-native-heap/ios/Vendor/libHeap.a(Heap.o), building for iOS Simulator, but linking in object file built for iOS, for architecture arm64

This commit prevents us from building for arm64, but this is mostly to unblock myself as I try to hammer out the remaining issues with TestDriver. A more permanent fix would look at the host operating system and determine which architecture to build for based on that.

(See https://stackoverflow.com/questions/63607158/xcode-12-building-for-ios-simulator-but-linking-in-object-file-built-for-ios.)

refs HEAP-20926